Moved event notes modal one folder up.
